Trait kernel_api::memory::allocator::BackingAllocator

pub unsafe trait BackingAllocator: Send + Sync {
    // Required methods
    fn allocate_contiguous(
        &self,
        frame_count: usize
    ) -> Result<Frame, AllocError>;
    unsafe fn deallocate_contiguous(
        &self,
        base: Frame,
        frame_count: NonZeroUsize
    );

    // Provided methods
    fn allocate_one(&self) -> Result<Frame, AllocError> { ... }
    fn try_allocate_zeroed(
        &self,
        frame_count: usize
    ) -> Result<Frame, ZeroAllocError> { ... }
    fn allocate_zeroed(&self, frame_count: usize) -> Result<Frame, AllocError> { ... }
    fn new(config: Config<'_>) -> Arc<dyn BackingAllocator>
       where Self: Sized { ... }
    fn push(&mut self, allocation: AllocationMeta) { ... }
    fn drain_into(self, into: &mut dyn BackingAllocator)
       where Self: Sized { ... }
    fn allocate_contiguous_aligned(
        &self,
        count: NonZeroUsize,
        alignment_log2: u32
    ) -> Result<Frame, AllocError> { ... }
    fn try_allocate_contiguous_aligned(
        &self,
        count: NonZeroUsize,
        alignment_log2: u32
    ) -> Result<Frame, AlignedAllocError> { ... }
}

An allocator that managed physical memory

In future, this may be replaced by a more general resource allocator. In that case, this trait will be deprecated and implementors will be expected to move to the more general trait.

Required Methods

fn allocate_contiguous(&self, frame_count: usize) -> Result<Frame, AllocError>

Allocates a contiguous range of physical memory

unsafe fn deallocate_contiguous(&self, base: Frame, frame_count: NonZeroUsize)

Safety

Must be deallocated with the same allocator that made the allocation

Provided Methods

fn allocate_one(&self) -> Result<Frame, AllocError>

Allocates a single Frame

fn try_allocate_zeroed( &self, frame_count: usize ) -> Result<Frame, ZeroAllocError>

🔬This is a nightly-only experimental API. (kernel_allocation_zeroing)

Tries to allocate a contiguous region of frame_count frames from a prezeroed buffer

fn allocate_zeroed(&self, frame_count: usize) -> Result<Frame, AllocError>

🔬This is a nightly-only experimental API. (kernel_allocation_zeroing)

Allocates a contiguous region of frame_count frames, manually zeroing them if there were no prezeroed frames

fn new(config: Config<'_>) -> Arc<dyn BackingAllocator>where Self: Sized,

🔬This is a nightly-only experimental API. (kernel_allocation_new)

fn push(&mut self, allocation: AllocationMeta)

🔬This is a nightly-only experimental API. (kernel_allocation_new)

fn drain_into(self, into: &mut dyn BackingAllocator)where Self: Sized,

🔬This is a nightly-only experimental API. (kernel_allocation_new)

fn allocate_contiguous_aligned( &self, count: NonZeroUsize, alignment_log2: u32 ) -> Result<Frame, AllocError>

🔬This is a nightly-only experimental API. (kernel_allocation_aligned)

Allocate a continuous range of count frames, aligned to 2^alignment_log2 frames

Errors

Returns an AllocError if the allocation could not be made

fn try_allocate_contiguous_aligned( &self, count: NonZeroUsize, alignment_log2: u32 ) -> Result<Frame, AlignedAllocError>

🔬This is a nightly-only experimental API. (kernel_allocation_aligned)

Allocate a continuous range of count frames. If the alignment of 2^alignment_log2 frames cannot be satisfied, make an allocation of the same size with a lower alignment. The number of allocated frames shall be rounded up to the returned alignment

Errors

Returns an AlignedAllocError if the originally requested alignment could not be satisfied. The new alignment shall be returned in AlignError.