Module kernel_api::sync

Provides kernel synchronisation primitives

These are currently based on spinlocks but this may be changed in future

Structs

LazyLockExperimental

OnceExperimental

OnceLockExperimental

Type Aliases

SpinlockExperimental

SpinlockGuardExperimental

Mutex

A mutual exclusion primitive useful for protecting shared data

MutexGuard

An RAII implementation of a “scoped lock” of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

RwLock

A reader-writer lock

RwReadGuard

RAII structure used to release the shared read access of a lock when dropped.

RwUpgradableReadGuard

RAII structure used to release upgradable read access of a lock when dropped.

RwWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped.